Implemented POST request endpoint, changed jobID from Integer to Long type which fixed API issues

+++ src/main/java/com/board/jobseeker/JobEntryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 // extend sub-interface of Spring Data's |Repository|, automatically generating CRUD methods
 // CrudRepository<JobEntry, Long> indicates to repository that domain type is JobEntry, with it's ID being a Long 
-interface JobEntryRepository extends CrudRepository<JobEntry, Integer> {
+interface JobEntryRepository extends CrudRepository<JobEntry, Long> {
 
 }
